Add files via upload
The files labeled S.I. are those which corresponds directly to the supporting information as mention in the article "Evaluating the environmental hazard of industrial chemicals from data collected during the REACH registration process" by B. Mikael Gustavsson, Andreas Hellhof and Thomas Backhaus. The table ecotoxicity data is, just as S.I. Table 1 a re-structuring of the data originally hosted by ECHA. However, due to the size of the database it is not fully curated. The ecotoxicity data is as complete as possible, but data given in running text within the original database is missing as they are impossible to capture. Use of the ecotoxicity data should thus be limited to hypothesis-generating as a guarantee on its completeness cannot be given.